cart.checkout method implemented for scenario conisting promotion for only one product

+++ src/main/java/com/example/shopping/processor/IPromoProcessor.java
@@ -8,6 +8,6 @@ import com.example.shopping.model.Product;
 
 public interface IPromoProcessor {
 	
-	public BigDecimal applyPromotion(List<Product> products);
+	public double applyPromotion(List<Product> products);
 
 }
